Make server and client each have their own logic component.
View component is passed a logic and it renders it (also sends input to
it).

This allows removing "if components.server == nil" special cases in
client code.

The current client-server-view mode truly simulates a full server and
full client (and renders using client's logic, not server's). This is
somewhat suboptimal for final production, and in the future can be
implemented with a server component only. Local player can be added
directly to server rather than via networking as a full blown client.