Implement player health, being hit, death logic.
DeadState is needed because we stop tracking the players state after
they die, so the interpolated position may become incorrect (it will
wrap around after some time passes).

An alternative future implementation is to track state of dead players
in the same way as alive players (and perhaps even idle ones). But only
if it can be made memory and bandwidth efficient.