Disable vertex attribute arrays after using program.
Enabling a vertex attribute array is a global operation, it's not bound
to a specific program.

Some programs have a different number of vertex attribute arrays, and
they were left enabled after said program is no longer used.

This is a problem for WebGL because of
https://www.khronos.org/registry/webgl/specs/1.0/#6.4.

The http://www.mjbshaw.com/2013/03/webgl-fixing-invalidoperation.html
article was very helpful in understanding the root issue.

Also check for GL error once per frame in order to catch any issues.

Fixes #12.